% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_dret_dw.R
\name{approx_dret_dw}
\alias{approx_dret_dw}
\title{Estimates derivative of portfolio volatility w.r.t. w.}
\usage{
approx_dret_dw(series, w, delta_val = 0.01, freq = 12)
}
\arguments{
\item{series}{xts return series.}

\item{w}{Portfolio weights.}

\item{delta_val}{weight variation.}

\item{freq}{times per year.}
}
\value{
Discrete derivative for each asset.
}
\description{
Estimates derivative of portfolio volatility w.r.t. w.
}
