% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_data}
\alias{asset_data}
\title{Informaci칩n de activos.}
\format{Data frame con informaci칩n de los activos.}
\source{
\url{http://www.bloomberg.info/}
}
\usage{
asset_data
}
\description{
Data frame con informaci칩n b치sica de activos disponibles.
}
\keyword{datasets}
