% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_universe.R
\name{asset_universe}
\alias{asset_universe}
\title{Asset Universe}
\usage{
asset_universe(mean_rets, cov_matrix, lambda)
}
\arguments{
\item{mean_rets}{Named vector of mean returns.}

\item{cov_matrix}{Covariance matrix.}

\item{lambda}{Penalty.}
}
\value{
Vector of asset names selected.
}
\description{
Estimates returns with a given period.
}
