% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash_conv.R
\name{cash_conv}
\alias{cash_conv}
\title{Cash converter.}
\usage{
cash_conv(cash_in, curr_in, spot, spot_id)
}
\arguments{
\item{cash_in}{Input amount.}

\item{curr_in}{Currency of input amount.}

\item{spot}{Spot rate.}

\item{spot_id}{Iso currency code.}
}
\value{
Converted amount.
}
\description{
Takes amount in some currency and converts ir to anoher currency.
}
