% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_index_series.R
\name{complete_index_series}
\alias{complete_index_series}
\title{Index Series}
\usage{
complete_index_series(series, backup_series)
}
\arguments{
\item{series}{Original series,}

\item{backup_series}{Series used to complete te original series.}
}
\value{
index series
}
\description{
Estimates portfolio mean return, variance and annualized volatility.
}
