% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_drawdown.R
\name{cond_drawdown}
\alias{cond_drawdown}
\title{Conditional Drawdown}
\usage{
cond_drawdown(series, w, pers_end_ind, pers_matrix, quant = 0.9,
  type = "log")
}
\arguments{
\item{series}{data frame series.}

\item{w}{Portfolio weights.}

\item{pers_end_ind}{Periods endpoint index.}

\item{pers_matrix}{Periods matrix. Used to speed up computations and parallel computing.}

\item{quant}{Quantile.}

\item{type}{Type of returns: arithmetic (discrete) or log (continuous)}
}
\value{
Conditional Drawdown.
}
\description{
Estimates de covariance using weighted averages of products of past returns.
}
