% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.R
\name{covar}
\alias{covar}
\title{Covariance matrix and volatilities.}
\usage{
covar(series, per = 12)
}
\arguments{
\item{series}{data frame with return series.}

\item{per}{returns period expressed as number of periods per year.}
}
\value{
Covariance matrix and volatility of each variable.
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
