% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_weight.R
\name{covar_weight}
\alias{covar_weight}
\title{Covariance matrix and volatilities using weighted averages.}
\usage{
covar_weight(series, weights, horizon)
}
\arguments{
\item{series}{data frame with return series.}

\item{weights}{Weights. It can be a declining function of time.}

\item{horizon}{Horizon}
}
\value{
Covariance matrix and volatility of each variable.
}
\description{
Estimates de covariance using weighted averages of products of past returns.
}
