% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdown.R
\name{drawdown}
\alias{drawdown}
\title{Drawdown Returns returns}
\usage{
drawdown(series, w, horizon = "12M", quant = 0.9, atribution = FALSE,
  type = "log")
}
\arguments{
\item{series}{data frame series.}

\item{w}{Portfolio weights.}

\item{horizon}{Drawdown horizon.}

\item{quant}{Quantile.}

\item{atribution}{Risk atribution}

\item{type}{Type of returns: arithmetic (discrete) or log (continuous)}
}
\value{
Drawdown distribution, and mean, max and conditional drawdown.
}
\description{
Estimates de covariance using weighted averages of products of past returns.
}
