% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_cross.R
\name{fx_cross}
\alias{fx_cross}
\title{fx_cross.}
\usage{
fx_cross(base_curr, ref_curr, curr_mkt_base, fx_base, curr_mkt_ref, fx_ref)
}
\arguments{
\item{base_curr}{Base currency.}

\item{ref_curr}{Reference currency.}

\item{curr_mkt_base}{Currency iso quote of base currency.}

\item{fx_base}{Base currency market fx rate.}

\item{curr_mkt_ref}{Currency iso quote of ref currency.}

\item{fx_ref}{Ref currency market fx rate.}
}
\value{
Cross fx rate.
}
\description{
Cross fx rate .
}
