% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedged_return.R
\name{hedged_return}
\alias{hedged_return}
\title{Hedged_return.}
\usage{
hedged_return(hedge, ret, fx_ret, fwd_prem, exp_ret = 0)
}
\arguments{
\item{hedge}{Hedged percentage.}

\item{ret}{Return in foreign (domectic) currency.}

\item{fx_ret}{Fx return.}

\item{fwd_prem}{Forward premium.}

\item{exp_ret}{Expected return over initial capital (1) over investment horizon.Given the difficulties to estimate the this value, it is assumed to be cero.}
}
\description{
Estimates hedged return.
}
