% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_return.R
\name{implied_return}
\alias{implied_return}
\title{Equilibrium returns}
\usage{
implied_return(lambda, Sigma, w, per = 12, rfr = 0)
}
\arguments{
\item{lambda}{risk aversion coefficient.}

\item{Sigma}{covariance matrix.}

\item{w}{Named vector of market capitalization weigths.}

\item{per}{returns period expressed as number of periods per year.}

\item{rfr}{Risk free rate conrresponding to the same period.}
}
\value{
Named vector of equilibrium returns. Annualized and not. The return is assumed to be continuous, i.e. estimated using natural log.
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
