% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_series.R
\name{index_series}
\alias{index_series}
\title{Index Series}
\usage{
index_series(series_list, weights, dates, val_ini = 100,
  invest_assets = NULL)
}
\arguments{
\item{series_list}{Portfolio weights.}

\item{weights}{named vector of weights.}

\item{dates}{Dates.}

\item{val_ini}{Initial value.}

\item{invest_assets}{Index, ETF.}
}
\value{
index series
}
\description{
Estimates portfolio mean return, variance and annualized volatility.
}
