% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_quote.R
\name{iso_quote}
\alias{iso_quote}
\title{ISO quote}
\usage{
iso_quote(curr1, curr2 = NULL, base_currs = c("NZD", "AUD", "EUR", "GBP"))
}
\arguments{
\item{curr1}{Currency 1.}

\item{curr2}{Currency 2.}

\item{base_currs}{Base currencies.}
}
\value{
iso currency quote.
}
\description{
Takes id of two currencies amd generates an iso Quote.
}
