% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market_cap}
\alias{market_cap}
\title{Capitalización de mercado.}
\format{Data frame con información de capitalización de mercado de los activos.}
\source{
\url{http://www.bloomberg.info/}
}
\usage{
market_cap
}
\description{
Información de capitalización de mercado en USD asociada a cada clase de activo.
}
\keyword{datasets}
