% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_drawdown.R
\name{max_drawdown}
\alias{max_drawdown}
\title{Max Drawdown Return}
\usage{
max_drawdown(per, w, series, type)
}
\arguments{
\item{per}{Dates period.}

\item{w}{Portfolio weights.}

\item{series}{data frame series.}

\item{type}{Type of returns: arithmetic (discrete) or log (continuous)}
}
\value{
Max Drawdown for particular period.
}
\description{
Estimates de covariance using weighted averages of products of past returns.
}
