% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_panel.R
\name{my_panel}
\alias{my_panel}
\title{my_panel}
\usage{
my_panel(x, y, z, ...)
}
\arguments{
\item{x}{Input x.}

\item{y}{Input y.}

\item{z}{Matrix data.}
}
\value{
None.
}
\description{
Auxliary function for lattice level plot.
}
