% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_rets_perc.R
\name{neg_rets_perc}
\alias{neg_rets_perc}
\title{Negative returns percentage in horizon.}
\usage{
neg_rets_perc(series, w, horizons = "12M")
}
\arguments{
\item{series}{data frame series.}

\item{w}{Portfolio weights.}

\item{horizons}{Vector of horizons in number of months.}

\item{type}{Type of returns: arithmetic (discrete) or log (continuous)}
}
\value{
Negative returns percentage in horizon.
}
\description{
Negative returns percentage in horizon. When counting negative returns it does not matter if returns are log or arthmetic.
}
