% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_portfolio.R
\name{optim_portfolio}
\alias{optim_portfolio}
\title{Optimal portfolio.}
\usage{
optim_portfolio(w_ini, fn, lb, ub, eqfun, eqB, w_bench = NULL,
  lb_act = NULL, ub_act = NULL, ineqfun = NULL, ineqLB = NULL,
  ineqUB = NULL, method = "RI", fixed = NULL, n.restarts = 10,
  n.sim = 20000, type = "absolute", itermax = 1000, NP = 100,
  max.time = 10)
}
\arguments{
\item{w_ini}{Initial weights.}

\item{fn}{Objective function or list of objective function. By default it corresponds to the mean-var utility.}

\item{lb}{Lower bound.}

\item{ub}{Upper bound.}

\item{eqfun}{Equality constraint function returning vector.}

\item{eqB}{Equality constraints.}

\item{w_bench}{Benchmark weigths.}

\item{lb_act}{Lower bound active weight per asset}

\item{ub_act}{Upper bound active weight per asset}

\item{ineqfun}{Inequality constraint function returning vector.}

\item{ineqLB}{Inequality lower bound.}

\item{ineqUB}{Inequality upper bound.}

\item{method}{Gradient descent (GD), GD with random initialization (RI), differential evolution (DE), Genetic Opt. using Derivative (GE), Generalized Simulated Annealing (SA) or Memetic with local search (MALS).}

\item{fixed}{Numeric index indicating parameters that stay fixed.}

\item{n.sim}{Random parameters for every restart of the solver.}

\item{type}{Type of objective function. absolute or relative.}

\item{itermax}{Maximum iteration (population generation) allowed.}

\item{NP}{Number of population members.}

\item{max.time}{Max time in seconds. Applied to SA.}

\item{n.restars}{Number of solver restarts.}
}
\value{
Optimal weights.
}
\description{
Optimal Portfolio.
}
