% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_contrib_uniroot.R
\name{optimal_contrib_uniroot}
\alias{optimal_contrib_uniroot}
\title{optimal contribution to reach target.}
\usage{
optimal_contrib_uniroot(pr_simul, port, rc_1 = 0.25, rc_2 = 0.333,
  rc_3 = 0.5, L_contrib = NA, L = 12, B = 2.5, Y = 0, gr_1, gr_2,
  pe_target = c(0.1, 0.2, 0.3))
}
\arguments{
\item{pr_simul}{Price Simul.}

\item{port}{Portfolio.}

\item{rc_1}{Contrib rate year 1.}

\item{rc_2}{Contrib rate year 2.}

\item{rc_3}{Contrib rate year 3 onwards.}

\item{L_contrib}{Contribution years.}

\item{L}{Fund Horizon}

\item{B}{Param distribution over time.}

\item{Y}{Yield.}

\item{gr_1}{Growth rate 2nd year onwards.}

\item{gr_2}{Growth rate 2nd year onwards.}

\item{pe_target}{PE target.}
}
\value{
optimal contribution to reach target.
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
