% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_simul.R
\name{params_simul}
\alias{params_simul}
\title{Mean vector and Covariance matrix.}
\usage{
params_simul(series, mu_in = NA, sigma_in = NA, period = "monthly")
}
\arguments{
\item{series}{data frame with series.}

\item{period}{Returns period .}

\item{asset_df}{Assets data frame with characteristics.}
}
\value{
Mean vector and Covariance matrix.
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
