% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_consol.R
\name{pe_consol}
\alias{pe_consol}
\title{PE Consol.}
\usage{
pe_consol(pe_simul)
}
\arguments{
\item{pe_simul}{PE simulation.}
}
\value{
PE Consol
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
