% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_t.R
\name{pe_t}
\alias{pe_t}
\title{PE at time ti.}
\usage{
pe_t(pe_simul, ti)
}
\arguments{
\item{pe_simul}{PE Simulation.}

\item{ti}{Time.}
}
\value{
PE at time ti.
}
\description{
PE at time ti.
}
