% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_limits.R
\name{port_limits}
\alias{port_limits}
\title{Estimates lower and upper limits of portfolio.}
\usage{
port_limits(port_series, series, w, port_vol_lim, perc_ret_var = 0.2,
  freq = 12, N = 1000, vol_factor = 0.1, min_max_q = c(0.975, 0.6),
  bias_bound_lim = 0.9, dw = 0.01)
}
\arguments{
\item{port_series}{xts portfolio return series.}

\item{series}{xts return series.}

\item{w}{Portfolio weights.}

\item{freq}{Returns period as times per year.}

\item{N}{Sample size to estimate vol. of each weight.}

\item{vol_factor}{Constant that multiplies portfolio volatility to perturbate portfolio returns for bootstrapping procedure.}

\item{min_max_q}{Min and max probability associated to min and max weights standard deviation.}

\item{bias_bound_lim}{If prob. of certain weight is larger than this param., then no bias-limit is imposed.}

\item{dw}{weight variation.}

\item{port_vol_ann_lim}{Anualized volatility portfolio limits.}

\item{sharpe}{User Sharpe or vol.}
}
\value{
lower and upper limits of portfolio.
}
\description{
Estimates lower and upper limits of portfolio.
}
