% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_return.R
\name{port_return}
\alias{port_return}
\title{Porfolio returns}
\usage{
port_return(series, weights, horizon)
}
\arguments{
\item{series}{data frame with return series.}

\item{weights}{Weights. It can be a declining function of time.}

\item{horizon}{Horizon}
}
\value{
Covariance matrix and volatility of each variable.
}
\description{
Estimates de covariance using weighted averages of products of past returns.
}
