% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_simul.R
\name{port_simul}
\alias{port_simul}
\title{Portfolio values.}
\usage{
port_simul(port_list, capital, afil_dist, afil_cf, series_start_per = 12)
}
\arguments{
\item{port_list}{portfolio list,}

\item{afil_dist}{Client portolio distribuion.}

\item{afil_cf}{Client cash flows.}

\item{series_start_per}{Series start period.}
}
\value{
Portfolio simulation.
}
\description{
Simulate portafolios values.
}
