% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_backtest.R
\name{portfolio_backtest}
\alias{portfolio_backtest}
\title{Porfolio backtest}
\usage{
portfolio_backtest(weights, capital, currency, asset_data, series_backtest,
  fx_hedge_asset = rep(0, length(weights)), fwd_prem = NULL,
  hold_per = "1M", rebal_per_in_months = NA, slippage = 5,
  commission = 5, invest_assets = NULL)
}
\arguments{
\item{weights}{Portfolio weights.}

\item{capital}{Initial capital in reference currency.}

\item{currency}{Reference currency.}

\item{asset_data}{Datafrane with info regarding each asset.}

\item{series_backtest}{xts series for backtest.}

\item{fx_hedge_asset}{Forward hedge ratio per asset.}

\item{fwd_prem}{forward premium corresponding to holding period (hold_per).}

\item{rebal_per_in_months}{Rebalancing period in months.}

\item{slippage}{Slippage basis points.}

\item{commission}{Commission basis points.}

\item{invest_assets}{Investable asset. By default: Index. It can be set to ETF or IA (investable asset).}
}
\value{
Backtesting results.
}
\description{
Estimates de covariance using weighted averages of products of past returns.
}
