% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_fx_hedge.R
\name{portfolio_fx_hedge}
\alias{portfolio_fx_hedge}
\title{Porfolio fx hedge}
\usage{
portfolio_fx_hedge(w, ref_curr, asset_data, series_list, series_fxfwd_list,
  dates, hold_per = "1M", exp_ret = 0, group.by = "Asset",
  bounded = TRUE)
}
\arguments{
\item{w}{Portfolio weights.}

\item{ref_curr}{Reference currency.}

\item{asset_data}{Datafrane with info regarding each asset.}

\item{series_list}{xts series (indexes and currencies).}

\item{series_fxfwd_list}{xts fx forward series. Outrights and forward premiums.}

\item{dates}{Initial and final dates.}

\item{hold_per}{Holding period return. 1M or 3M.}

\item{exp_ret}{Expected return over initial capital (1) over investment horizon.Given the difficulties to estimate the this value, it is assumed to be cero.}

\item{group.by}{Set of assets that share the same hedging ratio. By default the grouping is per currency. c("All", "Asset_Class", "Asset", "Currency", "Country")}

\item{bounded}{Hedge ratio bounded to the range 0-1.}
}
\value{
Fx hedge returns.
}
\description{
Estimates the static hedging ratio per group. See Schmittmann (2010).
}
