% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios_units_values.R
\name{portfolios_units_values}
\alias{portfolios_units_values}
\title{Portfolio unit values}
\usage{
portfolios_units_values(pr_simul, port, port_ini_cc, rc_1 = 0.25,
  rc_2 = 0.333, rc_3 = 0.5, L_contrib = NA, L = 12, B = 2.5, Y = 0,
  gr_1, gr_2, pe_target = c(0.1, 0.2, 0.3), delta_contrib = 0.2,
  pers_to_target = 1)
}
\arguments{
\item{pr_simul}{Price Simul.}

\item{port}{Portfolio.}

\item{port_ini_cc}{Initial Contributions}

\item{rc_1}{Contrib rate year 1.}

\item{rc_2}{Contrib rate year 2.}

\item{rc_3}{Contrib rate year 3 onwards.}

\item{L_contrib}{Contribution years.}

\item{L}{Fund Horizon}

\item{B}{Param distribution over time.}

\item{Y}{Yield.}

\item{gr_1}{Growth rate 2nd year onwards.}

\item{gr_2}{Growth rate 2nd year onwards.}

\item{pe_target}{PE target.}

\item{delta_contrib}{Contribution change.}

\item{pers_to_target}{Periods to reach targe.}
}
\value{
Portfolio unit values.
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
