% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_params.R
\name{posterior_params}
\alias{posterior_params}
\title{Black Litterman posterior mean and variance.}
\usage{
posterior_params(mu, q, tau, Sigma, P, Omega = NULL, conf = 0.5)
}
\arguments{
\item{mu}{Expected equilibrium return.}

\item{q}{Views relative and absolute returns.}

\item{tau}{Factor to reduce equilibrium returns variance.}

\item{Sigma}{Covariance matrix.}

\item{P}{Views matrix.}

\item{conf}{Confidence level on the view. Value between 0 and 1. This value is mapped to [0, 10] with 0.5 correponding to conf = 1.}

\item{omega}{Views variance matrix. Bu default: tau.diag(P.Sigma.t(P))/conf (Meucci (2009)).}
}
\value{
Returns.
}
\description{
Black Litterman posterior return incorporating views.
}
