% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_hw_simul.R
\name{pr_hw_simul}
\alias{pr_hw_simul}
\title{Simulated prices within the hull-white model,}
\usage{
pr_hw_simul(date_ini, id_swaps, cf_matrix, proy_leg1, proy_leg2, nom_leg1,
  nom_leg2, pr_leg1, pr_leg2, rate_leg1, rate_leg2, curr_leg1, curr_leg2,
  disc_leg1, disc_leg2, str_leg1, str_leg2, first_rate_float_leg1,
  first_rate_float_leg2, prev_cf_dates_leg1, prev_cf_dates_leg2,
  next_cf_dates_leg1, next_cf_dates_leg2, curves_list, curves_curr, spot_cop_id,
  spot_cop_array, spot_id, spot_array, rt_simul_list, simul_days, a, sigma,
  drift_fx, base = 360, M = 1000, base_currs = c("NZD", "AUD", "EUR",
  "GBP"), drift_adjust_usdcop = 0)
}
\arguments{
\item{w_ini}{Initial weights.}

\item{fn}{Objective function or list of objective function. By default it corresponds to the mean-var utility.}

\item{lb}{Lower bound.}

\item{ub}{Upper bound.}
}
\value{
Optimal weights.
}
\description{
HW simulation.
}
