% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_simul_trans.R
\name{pr_simul_trans}
\alias{pr_simul_trans}
\title{Mean vector and Covariance matrix.}
\usage{
pr_simul_trans(pr_simul_orig, asset_df, ref_curr = "COP", names_series,
  assets, normalize = TRUE)
}
\arguments{
\item{pr_simul_orig}{Simulated prices in original currency.}

\item{asset_df}{Assets data frame with characteristics.}

\item{ref_curr}{Reference currency.}

\item{names_series}{Series names.}

\item{normalize}{Normalize series.}
}
\value{
Simulated prices in ref currency.
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
