% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_equity_simul.R
\name{private_equity_simul}
\alias{private_equity_simul}
\title{PE simulation.}
\usage{
private_equity_simul(cc, M, V, rc_1 = 0.25, rc_2 = 0.333, rc_3 = 0.5,
  L_contrib = NA, L = 12, B = 2.5, Y = 0, gr_1, gr_2)
}
\arguments{
\item{cc}{Contribution.}

\item{M}{Number of simulations.}

\item{V}{Number of vintages.}

\item{rc_1}{Contrib rate year 1.}

\item{rc_2}{Contrib rate year 2.}

\item{rc_3}{Contrib rate year 3 onwards.}

\item{L_contrib}{Contribution years.}

\item{L}{Fund Horizon}

\item{B}{Param distribution over time.}

\item{Y}{Yield.}

\item{gr_1}{Growth rate 2nd year onwards.}

\item{gr_2}{Growth rate 2nd year onwards.}
}
\value{
PE simulation.
}
\description{
Takes the risk aversion coefficient, the covariance matrix and the market cap weights.
}
