% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{returns}
\alias{returns}
\title{Returns}
\usage{
returns(series, dates = NULL, period = "daily", type = "arithmetic")
}
\arguments{
\item{series}{object of state prices, or an OHLC type object.}

\item{dates}{Initial and end date.}

\item{period}{returns period. c(‘daily’, ‘weekly’, ‘monthly’, ‘quarterly’, ‘yearly’)}

\item{type}{type of returns: arithmetic (discrete) or log (continuous)}
}
\value{
Returns
}
\description{
Estimates returns with a given period.
}
