% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_atrib.R
\name{risk_atrib}
\alias{risk_atrib}
\title{Estimates risk attributions per factor for volatility, VaR and CVaR assuming normality.}
\usage{
risk_atrib(series, w, quant, normal = FALSE)
}
\arguments{
\item{series}{Portfolio weights.}

\item{w}{Portfolio weights.}

\item{quant}{Quantile}

\item{covar_mat}{Covariance matrix}
}
\value{
Losses distribution, VaR and CVaR.
}
\description{
Estimates risk attributions per factor for volatility, VaR and CVaR assuming normality..
}
