% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_fun.R
\name{risk_fun}
\alias{risk_fun}
\title{Risk functional}
\usage{
risk_fun(Sigma, type = "volatility", quant = 0.9)
}
\arguments{
\item{Sigma}{covariance matrix}

\item{type}{Type of risk measure. ('volatility','var', 'cvar')}

\item{quant}{Quantile.}
}
\value{
Risk function
}
\description{
Risk function.
}
