% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{series_list}
\alias{series_list}
\title{Series de tiempo.}
\format{Una lista donde cada componente corresponde a un objeto xts.}
\source{
\url{http://www.bloomberg.info/}
}
\usage{
series_list
}
\description{
Lista de series de objetos
}
\keyword{datasets}
