% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_merge.R
\name{series_merge}
\alias{series_merge}
\title{Series merge}
\usage{
series_merge(series_list, dates, asset_data, ref_curr, assets,
  currencies = NULL, convert_to_ref = FALSE, ref_per_unit_foreign = FALSE,
  invest_assets = NULL)
}
\arguments{
\item{series_list}{List of xts series.}

\item{dates}{Dates.}

\item{asset_data}{Datafrane with info regarding each asset.}

\item{ref_curr}{Id (iso) of reference currency.}

\item{assets}{Id of assets of interest.}

\item{convert_to_ref}{Indicator to define if series are converted to ref_curr.}

\item{ref_per_unit_foreign}{Indicator if currencies are converted to Base (ref) currency per unit of foreign currency.}

\item{invest_assets}{Investable asset. By default: Index. It can be set to ETF or IA (investable asset).}

\item{currecies}{Id (iso) of currencies of interest.}
}
\value{
xts series.
}
\description{
Series merge
}
