% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_var_cvar.R
\name{series_var_cvar}
\alias{series_var_cvar}
\title{Value at Risk and Conditional Value at risk.}
\usage{
series_var_cvar(series, quant, normal = FALSE)
}
\arguments{
\item{series}{returns series.}

\item{quant}{Quantile.}

\item{normal}{Indicator if returns are normaly distributed.}
}
\value{
Losses distribution, covar matrix, volatility, VaR and CVaR.
}
\description{
Estimates losses distribution, VaR amd CVaR.
}
