% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_simul.R
\name{state_simul}
\alias{state_simul}
\title{State simulation by Geometric Brownian Motion}
\usage{
state_simul(s0, times, mu, sigma, M)
}
\arguments{
\item{s0}{Initial state.}

\item{times}{Future times in years to simulate.}

\item{mu}{Drift.}

\item{sigma}{Volatility.}

\item{M}{Number of trajectories.}
}
\value{
Simulated trajectories.
}
\description{
Takes some initial state and simulates trajectories.
}
