% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_simul_mult.R
\name{state_simul_mult}
\alias{state_simul_mult}
\title{States simulation by Geometric Brownian Motion}
\usage{
state_simul_mult(S0, times, mu, sigma, rho, M = 1000)
}
\arguments{
\item{S0}{Vector of initial states.}

\item{times}{Future times in years to simulate.}

\item{mu}{Drift.}

\item{sigma}{Volatility.}

\item{rho}{Correlation matrix.}

\item{M}{Number of trajectories.}
}
\value{
Simulated trajectories.
}
\description{
Takes some initial states and smulates trajectories.
}
