% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_costs.R
\name{transaction_costs}
\alias{transaction_costs}
\title{Transaction costs}
\usage{
transaction_costs(units, price, slippage = 5, commission = 5)
}
\arguments{
\item{slippage}{Fixed percentage over asset price.}

\item{commission}{Percentage commission.}
}
\value{
Cost per transaction.
}
\description{
Basis points slippage, and BP commission per traded ammount. Future versions will contain other slippage models as we can have access to more data, e.g. trading volumes, number of trades.
}
