% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_ts_forecast.R
\name{uni_ts_forecast}
\alias{uni_ts_forecast}
\title{Univariate time series forecast}
\usage{
uni_ts_forecast(series, h)
}
\arguments{
\item{series}{returns series.}

\item{h}{number of steps ahead for forecast.}
}
\value{
Model order, coefficients and predictions.
}
\description{
Univariate time series forecast.
}
