% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fun.R
\name{utility_fun}
\alias{utility_fun}
\title{Objective function}
\usage{
utility_fun(type = "absolute", mu, Sigma, lambda, w_bench = NULL,
  min_var = FALSE, same_assets_bench = TRUE)
}
\arguments{
\item{type}{of objective function. absolute or relative.}

\item{mu}{Expected return}

\item{Sigma}{covariance matrix}

\item{lambda}{risk aversion coefficient}

\item{w_bench}{Benchmark weigths}

\item{min_var}{Indicator Min-Var portfolio}

\item{same_assets_bench}{Indicator function. If the portfolio and the benchmark assets are the same TRUE, otherwise FALSE.}
}
\value{
Objective function.
}
\description{
Utility function.
}
