% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fun_de.R
\name{utility_fun_de}
\alias{utility_fun_de}
\title{Objective function}
\usage{
utility_fun_de(type = "absolute", mu, Sigma, lambda, w_bench = NULL,
  lb = rep(0, length(mu)), ub = rep(1, length(mu)), lb_act = rep(0,
  length(mu)), ub_act = rep(1, length(mu)), min_var = FALSE,
  vol_obj = Inf, f_const = rep(1, length(mu)), f_const_lb = 0,
  f_const_ub = Inf)
}
\arguments{
\item{type}{Type of objective function. absolute or relative.}

\item{mu}{Expected return}

\item{Sigma}{covariance matrix}

\item{lambda}{risk aversion coefficient}

\item{w_bench}{Benchmark weigths}

\item{lb}{Lower bound per asset}

\item{ub}{Upper bound per asset}

\item{lb_act}{Lower bound active weight per asset}

\item{ub_act}{Upper bound active weight per asset}

\item{min_var}{Min-Var portfolio indicator}

\item{vol_obj}{Volatility objective. Default: Inf}

\item{f_const}{Matrix of factor constraints. Each entry is 0 or 1. Dimesion is NxM where N is the number of constraints and M is the number of assets.}

\item{f_const_lb}{Vector of lower bounds of each constraint.}

\item{f_const_ub}{Vector of upper bounds of each constraint.}
}
\value{
equilibrium returns. Annualized and not. The return is assumed to be continuous, i.e. estimated using natural log.
}
\description{
Utility function.
}
