% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_cvar.R
\name{var_cvar}
\alias{var_cvar}
\title{Value at Risk and Conditional Value at risk.}
\usage{
var_cvar(series, w, quant, normal = FALSE, port_name = "Portolio")
}
\arguments{
\item{series}{data frame with return series.}

\item{w}{Portfolio weights.}

\item{quant}{Quantile.}

\item{normal}{Indicator if returns are normaly distributed.}
}
\value{
Losses distribution, covar matrix, volatility, VaR and CVaR.
}
\description{
Estimates losses distribution, VaR amd CVaR.
}
