% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_universe.R
\name{asset_universe}
\alias{asset_universe}
\title{Asset Universe}
\usage{
asset_universe(mean_rets, cov_matrix, n_assets_univ, lambda = 0.1)
}
\arguments{
\item{mean_rets}{Named vector of mean returns. Can be implied returns.}

\item{cov_matrix}{(Annualized) Covariance matrix.}

\item{n_assets_univ}{Number of assets in the ivestable universe}

\item{lambda}{Correlation penalty.}
}
\value{
Vector of asset names selected. The names are in order, meaning the i position in the output vector is the i-th asset selected.
}
\description{
Suggests a set of assets.
}
