% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmarks}
\alias{benchmarks}
\title{Benchmarks.}
\format{Tabla de benchmarks.}
\source{
\url{http://www.bloomberg.info/}
}
\usage{
benchmarks
}
\description{
Dataframe de benchmarks.
}
\keyword{datasets}

