% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{efficient_frontier}
\alias{efficient_frontier}
\title{Efficient frontier.}
\usage{
efficient_frontier(mu, Sigma, lambda = seq(0.5, 3, 0.25))
}
\arguments{
\item{mu}{Expected return}

\item{Sigma}{Covariance matrix}

\item{lambda}{Vector of risk aversion coefficients}
}
\value{
Vectors or portfolio means and volatilities for different levels of risk aversion, and smooth spline object.
}
\description{
Efficient Frontier.
}
