% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_portfolio_resamp.R
\name{optim_portfolio_resamp}
\alias{optim_portfolio_resamp}
\title{Optimal portfolio.}
\usage{
\method{Default:}{GD}(mu, Sigma, lb, ub, w_ini, lambda = 1, N = 200,
  M = 1000, plot_ef = FALSE, spar = 0, ineqfun = NULL, ineqLB = NULL,
  ineqUB = NULL, method = "GD", n.restarts = 10, n.sim = 20000,
  conf_int = 0.9)
}
\arguments{
\item{mu}{Vector of mean returns.}

\item{Sigma}{Covariance matrix.}

\item{lb}{Lower bound.}

\item{ub}{Upper bound.}

\item{lambda}{Risk aversion coefficient.}
}
\value{
Optimal weights, mean resampled optimal weights, matrix of sampled weights.
}
\description{
Optimal Portfolio. Assuming returns
}
