% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fun.R
\name{utility_fun}
\alias{utility_fun}
\title{Objective function}
\usage{
utility_fun(type = "absolute", mu, Sigma, lambda, w_bench = NULL,
  min_var = FALSE)
}
\arguments{
\item{type}{of objective function. absolute or relative.}

\item{mu}{Expected return}

\item{Sigma}{covariance matrix}

\item{lambda}{risk aversion coefficient}

\item{w_bench}{Benchmark weigths}

\item{min_var}{Min-Var portfolio indicator}
}
\value{
equilibrium returns. Annualized and not. The return is assumed to be continuous, i.e. estimated using natural log.
}
\description{
Utility function.
}
